requirements.txt

# =========================
# path: requirements.txt
# =========================
transformers>=4.43.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.3.0
opencv-python>=4.9.0.80
pandas>=2.2.0
tqdm>=4.66.0
pytest>=8.0.0